LG-10517: Refactor LexisNexis client and response (#9462)
* delete unused code

* changelog: Internal, DocAuth, remove code that is not being used

* combine lexis_nexis_response and true_id_response
